requirements.txt
# Runtime dependencies
aiohttp>=3.8.0

# Development dependencies (for testing)
pytest>=7.0.0
pytest-homeassistant-custom-component>=0.13.0
